Faltaba un true cuando guarda la compra

+++ src/main/java/ar/edu/grupoesfera/cursospring/dao/UsuarioPizzaDao.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 
 public interface UsuarioPizzaDao {
-    ArrayList<Usuario_Pizza> obtenerPizzasCompradasPorUsuario(Long usuarioId);
+    ArrayList<Usuario_Pizza> obtenerPizzasCompradasPorUsuario(Integer usuarioId);
 
     ArrayList<Usuario_Pizza> obtenerPizzasCompradasPorUsuario24hs(Long id);
